requirements_langchain.txt
streamlit
langchain
langchain-community
anthropic
sentence-transformers
chromadb
python-dotenv
huggingface-hub